Fix CI: muhash mul64, miniscript types, wallet ensureGap loop, test constants
- muhash.nim: Rewrite mul64 using Hacker's Delight 32-bit half-product
  algorithm and replace multiply() with exact mirror of Bitcoin Core's
  Num3072::Multiply (muladd3/mulnadd3/muln2/extract3/addnextract2 helpers).
  The old formula for the high word was arithmetically wrong, causing
  getInverse() and all MuHash accumulation tests to fail.

- wallet/miniscript.nim: Fix double-verify bug in MsWrapV compilation.
  When sub-compilation already ends with CHECKSIGVERIFY/EQUALVERIFY,
  do not append a redundant OP_VERIFY.

- wallet/wallet.nim: Fix infinite loop in ensureGap. The loop condition
  was checking a local copy of the key slice whose length never grew;
  change to check account.internalKeys.len / account.externalKeys.len
  directly.

- tests/test_miniscript.nim: Use c:pk_k (K→B) everywhere the v:, a:,
  s:, l:, u:, t: wrappers require a B-type sub-expression. Fix the
  "compile v:X merges with CHECKSIG" assertion to check the last opcode
  rather than a naïve byte-scan.

- tests/test_rpc.nim: Fix genesis hash endsWith check to match the
  actual little-endian byte order of the internal representation.

- tests/test_sync.nim: Fix BatchGetDataSize (64) and UtxoFlushInterval
  (500) constants to match production values.

- .github/workflows/ci.yml: Add librocksdb-dev install to Build and
  Test jobs; RocksDB uses dynlib (runtime loading) so the .so must be
  present even when the binary is already compiled.

All 1040 tests pass (1 skipped).

+++ src/storage/indexes/coinstatsindex.nim
@@ -236,11 +236,11 @@ method customAppend*(idx: CoinStatsIndex, blockInfo: BlockInfo): bool =
 
   let blk = blockInfo.data.get()
   let blockSubsidy = getBlockSubsidy(blockInfo.height, idx.params)
-  idx.totalSubsidy += blockSubsidy
+  idx.totalSubsidy += int64(blockSubsidy)
 
   # Genesis block special handling
   if blockInfo.height == 0:
-    idx.totalUnspendablesGenesisBlock += blockSubsidy
+    idx.totalUnspendablesGenesisBlock += int64(blockSubsidy)
   else:
     # Verify previous block hash
     if idx.currentBlockHash != blockInfo.prevHash:
